+++ src/lib/fragments/AgGridEnterprise.react.js
@@ -2,10 +2,15 @@ import React from 'react';
 import {ModuleRegistry} from 'ag-grid-community';
 import {
     AllEnterpriseModule,
+    IntegratedChartsModule,
     LicenseManager,
     SparklinesModule,
 } from 'ag-grid-enterprise';
-import {AgChartsEnterpriseModule} from 'ag-charts-enterprise';
+import {AgChartsCommunityModule} from 'ag-charts-community';
+import {
+    AgChartsEnterpriseModule,
+    LicenseManager as AgChartsLicenseManager,
+} from 'ag-charts-enterprise';
 import MemoizedAgGrid, {propTypes} from './AgGrid.react';
 
 // Register all enterprise features
@@ -15,9 +20,24 @@ ModuleRegistry.registerModules([
 ]);
 
 export default function DashAgGridEnterprise(props) {
-    const {licenseKey} = props;
+    const {licenseKey, chartsLicenseKey, dashEnableCharts} = props;
     if (licenseKey) {
         LicenseManager.setLicenseKey(licenseKey);
+    }
+    if (dashEnableCharts) {
+        if (dashEnableCharts === 'enterprise') {
+            const effectiveChartsLicenseKey = chartsLicenseKey || licenseKey;
+            if (effectiveChartsLicenseKey) {
+                AgChartsLicenseManager.setLicenseKey(effectiveChartsLicenseKey);
+            }
+            ModuleRegistry.registerModules([
+                IntegratedChartsModule.with(AgChartsEnterpriseModule),
+            ]);
+        } else {
+            ModuleRegistry.registerModules([
+                IntegratedChartsModule.with(AgChartsCommunityModule),
+            ]);
+        }
     }
     return <MemoizedAgGrid {...props} />;
 }