+++ src/lib/utils/propCategories.js
@@ -343,6 +343,7 @@ export const PROPS_NOT_FOR_AG_GRID = [
     'setProps',
     'loading_state',
     'enableEnterpriseModules',
+    'dashEnableCharts',
     'parentState',
     'persistence',
     'persisted_props',