+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Links "DE#nnn" prior to version 2.0 point to the Dash Enterprise closed-source D
 ## [unreleased]
 ### Added
 - [#453](https://github.com/plotly/dash-ag-grid/pull/453) Test for changelog entry
+- [#448](https://github.com/plotly/dash-ag-grid/pull/448) Added support for AG-Charts (split out in v33 of AG Grid)
 
 ### Changed
 - [#452](https://github.com/plotly/dash-ag-grid/pull/452)